# bpo-44289: Keep argument file object's current position in tarfile.is_tarfile