--- a/Lib/tarfile.py
+++ b/Lib/tarfile.py
@@ -2484,7 +2484,9 @@ def is_tarfile(name):
     """
     try:
         if hasattr(name, "read"):
+            pos = name.tell()
             t = open(fileobj=name)
+            name.seek(pos)
         else:
             t = open(name)
         t.close()