Since Python 3.9 tarfile.is_tarfile accepts not only paths but also files and file-like objects #18090.

Verification if a file or file-like object is a tar file modifies file object's current position.

https://bugs.python.org/issue44289

<!-- issue-number: [bpo-44289](https://bugs.python.org/issue44289) -->
https://bugs.python.org/issue44289
<!-- /issue-number -->