@akulakov 2021-06-17
@amateja hi, you need to sign the CLA..

Other than that, I've looked at the patch and it looks good to me. I've also confirmed that the fix does what it's supposed to by running test with & w/out the fix.

Since this is a (minor) backwards compatibility change, I'm not sure if it needs to also be added to "what's new", I hope @ethanfurman  can comment.