@amateja 2021-06-21
Hi @akulakov, thanks for reviewing.

I believe I've already signed CLA. I confirmed that with https://check-python-cla.herokuapp.com/.

Regarding "what's new" I entirely base on your judgment. Just let me know what should be done to unblock this change from being accepted.