@akulakov 2021-12-09
I looked at it again and I think this should be fixed in `open()` rather than in `is_tarfile()`, because is_tarfile uses open() so it should be fixed at the source; and because there's already logic in open() that partly addresses this issue:

https://github.com/python/cpython/blob/9fe0de28bcc8d6b4077d9c3994648049730ee721/Lib/tarfile.py#L1616

So the fix would be (a few lines above the linked line; I tested it & it works):

```
                    fp = func(name, "r", fileobj, **kwargs)
                    if fileobj is not None:
                        fileobj.seek(saved_pos)
                    return fp
```