@akulakov 2021-12-09
Note also that the fix I suggested covers more issues: it also covers calling `open()` more than a single time, e.g.:

    kwargs = {'fileobj' if hasattr(archive, 'read') else 'name': archive}
    t = tarfile.open(**kwargs)
    t = tarfile.open(**kwargs)
    return [member.name for member in t.getmembers()]