@amateja 2021-12-09
Thanks for your review @akulakov. You are right that this should be fixed at the source. I focused myself on `is_tarfile()` because new functionality for file-like objects handling was introduced only there with #18090. But I'm happy to change this approach.