@akulakov 2021-12-10
Thanks for making the changes!

I will do a bit of testing tomorrow and look over it again; I think it's enough to have the test that checks that open keeps the same position as you have it, because both bugs are effectively covered by this test.

I will try to post an update tomorrow or in the next 2 days.