@akulakov 2021-12-11
I did some profiling to make sure an extra seek doesn't affect performance, I did it by opening 1000 small tar files; here's what I got:

with patch: 415, 417, 418 msec per loop
without patch: 414, 419, 417 msec

As expected, it doesn't make noticeable difference.

```
2850 (~/temp2/temp) % ~/opensource/cpython2/python.exe -m timeit -c 'import tarfile' 'for x in range(1000): tarfile.open("./b%s.tar"%x)'
1 loop, best of 5: 415 msec per loop
2850 (~/temp2/temp) % ~/opensource/cpython2/python.exe -m timeit -c 'import tarfile' 'for x in range(1000): tarfile.open("./b%s.tar"%x)'
1 loop, best of 5: 417 msec per loop
2850 (~/temp2/temp) % ~/opensource/cpython2/python.exe -m timeit -c 'import tarfile' 'for x in range(1000): tarfile.open("./b%s.tar"%x)'
1 loop, best of 5: 418 msec per loop
2850 (~/temp2/temp) % r                                      --INS--
~/opensource/cpython2/python.exe -m timeit -c 'import tarfile' 'for x in range(1000): tarfile.open("./b%s.tar"%x)'
1 loop, best of 5: 414 msec per loop
2852 (~/temp2/temp) % r                                      --INS--
~/opensource/cpython2/python.exe -m timeit -c 'import tarfile' 'for x in range(1000): tarfile.open("./b%s.tar"%x)'
1 loop, best of 5: 419 msec per loop
2854 (~/temp2/temp) % r                                      --INS--
~/opensource/cpython2/python.exe -m timeit -c 'import tarfile' 'for x in range(1000): tarfile.open("./b%s.tar"%x)'
1 loop, best of 5: 417 msec per loop
```