@akulakov 2021-12-11
Please update the PR title and the news entry.

For the unit tests, keep both of them for now. A core developer will need to review the PR, so I would wait for them to make this call.