@akulakov 2021-12-11
I'm not a bit concerned that users may be relying on the initial seek of `open()`. I will read about the tar spec, perhaps there's a clean way to keep the initial seek but leave it for additional open()'s.