@akulakov 2021-12-12
I do think it's plausible that users may be relying on `open()` seeking to 512 bytes, so to minimize the backwards incompatibility change, it's best to allow the first open() to go to 512 bytes, but subsequent open() will reset to 0 first and then go to 512 again.

If users are relying on multiple opens to seek further, they are relying on buggy behaviour  and IMO it's ok to fix that.

Note that when open fails, we still always want to restore original position. It's only on success that  special casing of 512 position comes into play.

I did some testing and the following logic works:
```
                try:
                    if fileobj is not None and saved_pos==BLOCKSIZE:
                        fileobj.seek(0)
                    return func(name, "r", fileobj, **kwargs)
```

It's a bit ugly to hardcode the initial seek position like this, it may be someone more familiar with module internals will suggest a better fix.