@akulakov 2022-01-22
Andrzej: I have some spare time right now, if you prefer I can finish the PR and list you as a co-author (because I would have to make it into a new PR). If you are still planning to finish this, no rush.