@amateja 2022-01-24
I'm more than happy to work further on this case. I thought we are waiting for someone more familiar with tar internals to comment on your proposal.